Fix up wav upload issue and clean things up a bit/organize.
