+++ apps/piptris/assets/piptris.png
