+++ apps/piptris/install.html
@@ -1,11 +1,11 @@
 <!doctype html>
 <html lang="en">
   <head>
-    <link rel="stylesheet" href="../../css/spectre.min.css" />
-    <link rel="stylesheet" href="../../css/main.css" />
-    <link rel="stylesheet" href="../../css/main-overrides.css" />
-    <link rel="stylesheet" href="../../apps/piptris/piptris.css" />
-    <script src="../../core/lib/customize.js"></script>
+    <link rel="stylesheet" href="../../../css/spectre.min.css" />
+    <link rel="stylesheet" href="../../../css/main.css" />
+    <link rel="stylesheet" href="../../../css/main-overrides.css" />
+    <link rel="stylesheet" href="assets/web/piptris.css" />
+    <script src="../../../core/lib/customize.js"></script>
   </head>
   <body>
     <h2>Installer</h2>
@@ -52,6 +52,8 @@
     </div>
     <hr />
     <button id="installBtn" class="btn btn-primary" disabled>Install</button>
-    <script src="install.js"></script>
+    <script src="assets/web/install-log.js"></script>
+    <script src="assets/web/install.js"></script>
+    <script src="assets/web/preview-songs.js"></script>
   </body>
 </html>
