+++ apps/piptris/metadata.json
@@ -1,18 +1,21 @@
 {
   "id": "piptris",
   "name": "Piptris",
-  "version": "2.5.1",
+  "version": "2.5.2",
   "description": "Stack, rotate, and clear falling blocks in this Tetris inspired mini-game. Simple, addictive, and perfect for idle moments between scavenging runs or stimpak breaks.",
-  "icon": "piptris.png",
+  "icon": "assets/piptris.png",
   "tags": "game",
   "type": "app",
   "readme": "README.md",
   "custom": "install.html",
   "storage": [
     { "name": "USER/piptris.js", "url": "app.min.js" },
-    { "name": "USER/PIPTRIS/piptris.json", "url": "piptris.json" },
-    { "name": "USER/PIPTRIS/gear.json", "url": "gear.json" },
-    { "name": "USER/PIPTRIS/nuke.json", "url": "nuke.json" },
-    { "name": "USER/PIPTRIS/radioactive.json", "url": "radioactive.json" }
+    { "name": "USER/PIPTRIS/piptris.json", "url": "assets/piptris.json" },
+    { "name": "USER/PIPTRIS/gear.json", "url": "assets/gear.json" },
+    { "name": "USER/PIPTRIS/nuke.json", "url": "assets/nuke.json" },
+    {
+      "name": "USER/PIPTRIS/radioactive.json",
+      "url": "assets/radioactive.json"
+    }
   ]
 }
