+++ index.html
@@ -185,6 +185,7 @@
         </div>
       </div>
     </div>
+    <!-- CORE ESPRUINO SCRIPTS -->
     <script src="webtools/uart.js"></script>
     <script src="core/js/utils.js"></script>
     <script src="webtools/heatshrink.js"></script>
@@ -195,6 +196,8 @@
     <script src="core/js/comms.js"></script>
     <script src="core/js/appinfo.js"></script>
     <script src="core/js/index.js"></script>
+    <!-- ADDITIONS / OVERRIDES -->
+    <script src="scripts/buttons.js"></script>
     <script src="scripts/loader-overrides.js"></script>
   </body>
 </html>
